% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFR_plot.R
\name{CFR_plot}
\alias{CFR_plot}
\title{Plotting of CFR curves}
\usage{
CFR_plot (CFR,time,xmin,xmax,ymax=1,lab.x="Study time",lab.y="CFR",col=ifelse(is.vector(CFR),1,c(1:ncol(CFR))),cex.lab=1.5, cex.axis=1.5,lwd=ifelse(is.vector(CFR),1,c(1:ncol(CFR))),lty=ifelse(is.vector(CFR),1,c(1:ncol(CFR))))
}
\arguments{
\item{CFR}{Vector or data frame, CFR values for different methods}

\item{time}{time Points of interest}

\item{xmin}{Minimum value of the horizontal coordinate}

\item{xmax}{Maximum value of horizontal coordinate}

\item{ymax}{Maximum value of vertical coordinate}

\item{lab.x}{Horizontal Label}

\item{lab.y}{Vertical Label}

\item{col}{"col" parameters in function plot}

\item{cex.lab}{"cex.lab" parameters in function plot}

\item{cex.axis}{"cex.axis" parameters in function plot}

\item{lwd}{"cex.axis" parameters in function plot}

\item{lty}{"cex.axis" parameters in function plot}
}
\description{
Plotting of CFR curves
}
\examples{
###single method
Yip_CFR<-c(0.198,0.048,0.531)
time<-c(60,80,93)
CFR_plot(Yip_CFR,time,60,100)
###Multiple methods
CFRs<-data.frame(Chen=c(0.433,0.467,0.777),Yoshikura=c(0.157,0.149,0.122),Yip.cfr=c(0.198,0.048,0.531),Garske=c(0.246,0.252,0.298),Nishiura=c(0.308,0.312,0.318),Ejima=c(0.108,0.098,0.039))
time<-c(60,80,93)
CFR_plot(CFRs,time,60,100)
}
