% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chdata.R
\name{chdata}
\alias{chdata}
\title{Title Transform Individual Data By Survival Time}
\usage{
chdata(data)
}
\arguments{
\item{data}{individual survival data transformed by surdata}
}
\description{
Transform Individual Data By Survival Time.
}
\examples{
data(individual)
#get data
data1<-surdata(data=individual,t=60,type=1)
#study time:60;type:1,cout time.
chdata(data=data1)
}
