% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chen.cfr.R
\name{chen.cfr}
\alias{chen.cfr}
\title{Calculate CFR In Individual Survival Data}
\usage{
chen.cfr(data)
}
\arguments{
\item{data}{individual survival data transformed by surdata}
}
\value{
A list containing the following components:
\item{method}{the source}
\item{estimation}{ estimate of CFR and variance}
}
\description{
Calculate CFR using a constant cure-death hazard ratio,  See references for details
}
\examples{
data(individual)
data1<-surdata(data=individual,t=60,type=1)
chen.cfr(data1)
}
\references{
Zheng Chen. Estimating the case fatality rate using a constant cure-death hazard ration. Lifetime Data Anal. 2009,15,316-329.
}
