% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chen.R
\name{Chen}
\alias{Chen}
\title{Chen-Nakamura method}
\usage{
Chen(data,t)
}
\arguments{
\item{data}{individual survival data transformed by sumdata}

\item{t}{the study time}
}
\value{
A list containing the following components:
\item{CFR}{CFR}
\item{lower}{the lower of CFR's CI}
\item{upper}{the upper of CFR's CI}
\item{var}{the variance of CFR}
}
\description{
Chen-Nakamura method
}
\examples{
data(individual)
data1<-sumdata(individual)
Chen(data1,60)
}
\references{
Zheng Chen. Estimating the case fatality rate using a constant cure-death hazard ration. Lifetime Data Anal. 2009,15,316-329.
}
