% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ejima.R
\name{Ejima}
\alias{Ejima}
\title{Ejima method}
\usage{
Ejima(data,t,m,sd,distr=1)
}
\arguments{
\item{data}{summarized survival data}

\item{t}{the study time.}

\item{m}{Mean survival time of dead individuals}

\item{sd}{Variance of survival time of dead individuals}

\item{distr}{Distribution of survival time of dead individuals}
}
\value{
A list containing the following components:
\item{CFR}{CFR}
\item{lower}{the lower of CFR's CI}
\item{upper}{the upper of CFR's CI}
\item{var}{the variance of CFR}
}
\description{
Ejima method
}
\examples{
data(individual)
data1<-sumdata(individual)
Ejima(data1,60,25.67,71.74,1)
}
