% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Garske.R
\name{Garske}
\alias{Garske}
\title{Garske method}
\usage{
Garske(data,t,m,sd,distr)
}
\arguments{
\item{data}{summarized survival data}

\item{t}{the study time.}

\item{m}{Mean survival time of dead individuals}

\item{sd}{Variance of survival time of dead individuals}

\item{distr}{Distribution of survival time of dead individuals}
}
\value{
A list containing the following components:
\item{CFR}{CFR}
\item{lower}{the lower of CFR's CI}
\item{upper}{the upper of CFR's CI}
\item{var}{the variance of CFR}
}
\description{
Garske method
}
\examples{
data(individual)
data1<-sumdata(individual)
Garske(data1,60,25.67,71.74,1)
}
