% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.donkm.R
\name{out.donkm}
\alias{out.donkm}
\title{Title Calculating CFR Using KM method In Individual Data}
\usage{
out.donkm(data, p)
}
\arguments{
\item{data}{individual survival data transformed by surdata.}

\item{p}{starting parameter values. p[1:2] for the mean and variance of death,p[3:4] for the mean and variance of cure,p[5] for CFR.}
}
\value{
method the source.

estimation the estimate at the study time,which contains Time(study time),CFR,Var(variance) and 95%CI.

p parameter values

Var variance matrix
}
\description{
Calculating CFR using KM method in individual survival data. See references for details.
}
\references{
Christl A Donnelly. Epidemiological determinants of spread of causal agent of severe acute respiratory syndrome in Hong Kong. The Lancet,2003,361:1761-1832.
}
