% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepcov.p.R
\name{stepcov.p}
\alias{stepcov.p}
\title{Screening for statistically significant covariates using BICcr}
\usage{
stepcov.p(cov, p, data)
}
\arguments{
\item{cov}{cov covariates}

\item{p}{starting parameter values;p[1] for beta0,p[2:(n+1)] for beta1 with respect to death,p[(n+2):(2*n+1)] for beta2 with respect to cure.}

\item{data}{individual survival data transformed by surdata.}
}
\value{
A list containing the following components:
\item{BICcr}{BICcr value}
\item{method}{the source.}
\item{cov}{ covariate.}
\item{estimate}{ coefficient values.}
\item{var}{ variance matrix}
\item{time}{ study time}
\item{deaths}{ number of death}
}
\description{
Screening for statistically significant covariates using BICcr
}
\examples{
data(individual)
head(individual)
data1<-surdata(data=individual,t=60,type=1)
obj<-stepcov.p(data1,cov=c("sex","age"),p=c(0.2,0.4,0.5,0.4,0.6))
}
