% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumdata.R
\name{sumdata}
\alias{sumdata}
\title{Title}
\usage{
sumdata(data)
}
\arguments{
\item{data}{simulation data,for example sdata.gamma and sdata.wei}
}
\description{
Conversion of individual data to aggregated data
}
