% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surdata.R
\name{surdata}
\alias{surdata}
\title{Title Transform Individual Data for Calculation}
\usage{
surdata(data, t, type)
}
\arguments{
\item{data}{the collected individual data.}

\item{t}{the study time.}

\item{type}{the type of collected data.type=1 for cout time(e.g. 1,2....), type=2 for calendar time, type=3 for survival time.}
}
\value{
The returned data.frame contains original information only prior to cut point time(study time).
}
\description{
Transform collected individual survival data into which can be used to calculate CFR.
}
\examples{
data(individual)
#get data
surdata(data=individual,t=60,type=1)
#study time:60;type:1,cout time.
}
