% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yoshikura.R
\name{Yoshikura}
\alias{Yoshikura}
\title{Yoshikura Method}
\usage{
Yoshikura(data)
}
\arguments{
\item{data}{individual survival data transformed by sumdata}

\item{t}{the study time}
}
\value{
A list containing the following components:
\item{CFR}{CFR}
\item{lower}{the lower of CFR's CI}
\item{upper}{the upper of CFR's CI}
\item{var}{the variance of CFR}
}
\description{
Yoshikura Method
}
\examples{
data(individual)
data1<-sumdata(individual)
Yoshikura(data1,60)
}
