% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFR_plot.R
\name{CFR_plot}
\alias{CFR_plot}
\title{Plotting of CFR curves}
\usage{
CFR_plot (data,method,time=data$day,m=NULL,sd=NULL,distr=NULL,xmin,xmax,ymax=1,lab.x="Study time",lab.y="CFR",col=c(1:length(method)),cex.lab=1.5, cex.axis=1.5,lwd=c(1:length(method)),lty=c(1:length(method)))
}
\arguments{
\item{data}{summarized survival data}

\item{method}{Different methods of calculating the CFR. specifically "Yip","Chen","Garske","Nishiura","Ejima"}

\item{time}{time Points of interest}

\item{m}{Mean survival time of dead individuals. Methods "Garske", "Nishiura" and "Ejima"are required.}

\item{sd}{Variance of survival time of dead individuals.Methods "Garske", "Nishiura" and "Ejima"are required.}

\item{distr}{Distribution of survival time of dead individuals.Methods "Garske", "Nishiura" and "Ejima"are required.}

\item{xmin}{Minimum value of the horizontal coordinate}

\item{xmax}{Maximum value of horizontal coordinate}

\item{ymax}{Maximum value of vertical coordinate}

\item{lab.x}{Horizontal Label}

\item{lab.y}{Vertical Label}

\item{col}{"col" parameters in function plot}

\item{cex.lab}{"cex.lab" parameters in function plot}

\item{cex.axis}{"cex.axis" parameters in function plot}

\item{lwd}{"cex.axis" parameters in function plot}

\item{lty}{"cex.axis" parameters in function plot}
}
\description{
Plotting of CFR curves
}
\examples{
data(individual)
data1<-sumdata(individual)
CFR_plot(data=datasum,method=c("Yip","Chen","Garske"),m=25.67,sd=71.74,distr=2,xmin=0,xmax=100)
CFR_plot(data=datasum,method=c("Yip","Chen"),xmin=0,xmax=100)
}
