% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumdata.R
\name{sumdata}
\alias{sumdata}
\title{Aggregated data}
\usage{
sumdata(data)
}
\arguments{
\item{data}{simulation data,for example sdata.gamma and sdata.wei}
}
\value{
The returned data.frame on aggregated data
}
\description{
Conversion of individual data to aggregated data
}
\examples{
data(individual)
data1<-sumdata(individual)
}
